pyproject.toml
[project]
name = "inverse-kinematics-hexapod"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.8",
    "pyqt6>=6.11.0",
    "pyyaml>=6.0.3",
    "types-pyyaml>=6.0.12.20260408",
    "websockets>=16.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hexapod"]

[tool.pyright]
extraPaths = ["src"]

